Refactored the Content component.
See, it's the very next commit. I didn't have to change too much. Or,
anything for that matter, it's just VS Code. There are also a few
styling/margin changes, but nothing too serious.
