+++ src/Components/Layout/Content/responsive.css
@@ -1,6 +1,7 @@
 @media screen and (max-width: 1279px) {
     #main-content {
         margin: 0vh 2.5vw;
+        margin-top: 2.5vh;
     }
 
     /* This is just an easy way to put the navigation at the bottom. */
@@ -13,6 +14,7 @@
 @media screen and (min-width: 1280px) {
     #main-content {
         margin: 0 10vw;
+        margin-top: 2.5vh;
     }
 
     .content {
